Move str_format to str.cpp
